src/main/java/com/indorse/atm/locations/AtmLocationApplication.java
package com.indorse.atm.locations;

import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.boot.builder.SpringApplicationBuilder;
import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;

@SpringBootApplication
public class AtmLocationApplication extends SpringBootServletInitializer {

    public static void main(String[] args) {
        SpringApplication.run(AtmLocationApplication.class, args);
    }

    @Override
    protected SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
        return applicationBuilder.sources(AtmLocationApplication.class);
    }
}
